requirements.txt
requests
python-whois
dnspython
colorama
tqdm
python-dotenv